Fixed comments and change int to float.

+++ FIXED_BATCH_3/no_duplicate_numbers.py
@@ -1,15 +1,15 @@
-#Create a empty list
+# Create a empty list
 number_list = []
 
-#Use for loop to allow users to input 10 numbers
+# Use for loop to allow users to input 10 numbers
 for i in range(10):
-    number = int(input(f"[{i + 1}] Enter the number: "))
+    number = float(input(f"[{i + 1}] Enter the number: "))
 
-    #Check if the number are not in the list and add them
+    # Check if the number are not in the list and add them
     if number not in number_list:
         number_list.append(number)
 
-#Print result
+# Print result
 print(f"Result: {number_list}")
